+++ pkg/R/utils.R
@@ -23,9 +23,16 @@
 #' When you define a handler, you specify its name and the associated action.
 #' A few predefined actions described below are provided.
 #'
-#' A handler is a function that accepts a \var{logging.record} and handler
+#' A handler action is a function that accepts a formated message and handler
 #' configuration.
 #'
+#' Messages passed are filtered already regarding loglevel.
+#'
+#' \dots parameters are used by logging system to interact with the action. \dots can
+#' contain \var{dry} key to inform action that it meant to initialize itself. In the case
+#' action should return TRUE if initialization succeded.
+#'
+#' If it's not a dry run \dots contain the whole preformated \var{logging.record}.
 #' A \var{logging.record} is a named list and has following structure:
 #' \describe{
 #'   \item{msg}{contains the real formatted message}
@@ -34,15 +41,10 @@
 #'   \item{logger}{name of the logger that generated it}
 #'   \item{timestamp}{formatted message timestamp}
 #' }
-#' Messages passed are filtered already regarding loglevel.
 #'
-#' \dots parameters are used by logging system to interact with the action. \dots can
-#' contain \var{dry} key to inform action that it meant to initialize itself. In the case
-#' action sould return TRUE if initialization succeded.
-#'
-#' @param msg A \var{logging.record} to handle
-#' @param handler The handler environment containing its options. You can register the
-#'   same action to handlers with different properties.
+#' @param msg A formated message to handle.
+#' @param handler The handler environment containing its options. You can
+#'   register the same action to handlers with different properties.
 #' @param ... parameters provided by logger system to interact with the action.
 #'
 #' @examples
@@ -54,13 +56,65 @@
 NULL
 
 #' @rdname inbuilt-actions
+#'
+#' @details
+#' \code{writeToConsole} detects if crayon package is available and uses it
+#' to color messages. The coloring can be switched off by means of configuring
+#' the handler with \var{color_output} option set to FALSE.
+#'
 #' @export
 #'
 writeToConsole <- function(msg, handler, ...) {
-  if (length(list(...)) && "dry" %in% names(list(...)))
+  if (length(list(...)) && "dry" %in% names(list(...))) {
+    if (!is.null(handler$color_output) && handler$color_output == FALSE) {
+      handler$color_msg <- function(msg, level_name) msg
+    } else {
+      handler$color_msg <- .build_msg_coloring()
+    }
     return(TRUE)
+  }
+
+  stopifnot(length(list(...)) > 0)
+
+  level_name <- list(...)[[1]]$levelname
+  msg <- handler$color_msg(msg, level_name)
   cat(paste0(msg, "\n"))
 }
+
+.build_msg_coloring <- function() {
+  crayon_env <- tryCatch(asNamespace("crayon"),
+                         error = function(e) NULL)
+
+  default_color_msg <- function(msg, level_name) msg
+  if (is.null(crayon_env)) {
+    return(default_color_msg)
+  }
+
+  if (is.null(crayon_env$make_style) ||
+      is.null(crayon_env$combine_styles) ||
+      is.null(crayon_env$reset)) {
+    return(default_color_msg)
+  }
+
+  color_msg <- function(msg, level_name) {
+    style <- switch(level_name,
+                    "FINEST" = crayon_env$make_style("gray80"),
+                    "FINER" = crayon_env$make_style("gray60"),
+                    "FINE" = crayon_env$make_style("gray60"),
+                    "DEBUG" = crayon_env$make_style('deepskyblue4'),
+                    "INFO" = crayon_env$reset,
+                    "WARNING" = crayon_env$make_style('darkorange'),
+                    "ERROR" = crayon_env$make_style('red4'),
+                    "CRITICAL" =
+                      crayon_env$combine_styles(crayon_env$bold,
+                                                crayon_env$make_style('red1')),
+                    crayon_env$make_style("gray100"))
+    res <- paste0(style(msg), crayon_env$reset(""))
+    return(res)
+  }
+  return(color_msg)
+}
+
 
 #' @rdname inbuilt-actions
 #'