+++ pkg/NEWS.md
@@ -2,6 +2,7 @@
   * Documentation generation migrated to roxygen2
   * Tests migrated to testthat, test coverage enhanced
   * Most issues reported by goodpractice fixed
+  * If crayon is available console messages are in color
   
 # logging 0.8-104 (2018-11-29)
   * Maintener changed to Walerian Sokolowski